Fixed the priority issue. So now servers actually have priorities that could be set.
I solved the priority issue where the simple chat client couldn't find the servers in a good time and it kept finding the most recent server. I fixed this by having a threading event and that threading event calls set which sets a flag and tells the browser that we have successfully found a service. So basically, the main browser thread waits for the add_service thread to set the flag that says I found the best priority url.